+++ src/js/contact.js
@@ -1,19 +1,6 @@
+import nav from './nav';
+
 export default function contact() {
-  const nav = document.createElement('nav');
-  nav.classList.add('nav');
-  const ul = document.createElement('ul');
-  ul.classList.add('ul');
-
-  const links = ['Home', 'Menu', 'Contact'];
-  for (let i = 0; i < 3; i += 1) {
-    const li = document.createElement('li');
-    li.innerHTML = links[i];
-    li.classList.add('li');
-    ul.appendChild(li);
-  }
-
-  nav.appendChild(ul);
-
   const mapouter = document.createElement('div');
   mapouter.classList.add('mapouter');
 
@@ -107,9 +94,9 @@ export default function contact() {
 
   const h = document.createElement('div');
   h.classList.add('h');
-  h.appendChild(nav);
   h.appendChild(mapouter);
   h.appendChild(info);
 
   document.body.appendChild(h);
+  nav();
 }
